Create sync
This can be used on Unix-like (including macOS) or Linux systems with Rsync installed.

SSH for remote must first be configured, with addition of the client's public key to the remote host's authorized_keys (or similar).

On macOS, a passphrase is recommended for added security. For ease of use, the passphrase should be added to Keychain (ssh-add on Terminal).